Refactor the plugin list into its own component
