+++ openc3-cosmos-init/plugins/packages/openc3-vue-common/src/tools/admin/tabs/plugins/index.js
@@ -16,6 +16,7 @@
 # if purchased from OpenC3, Inc.
 */
 
+import PluginList from './PluginList.vue'
 import PluginListItem from './PluginListItem.vue'
 
-export { PluginListItem }
+export { PluginList, PluginListItem }
